requirements.txt
click
python-slugify